maajic

+++ libraries/main/MasterBT.cpp
@@ -21,7 +21,7 @@ void MasterBT::float2Bytes(float floatVal, byte * bytes_array){
 }
 
 void sendCoords(SensorGPS * currentGPS) {
-  float2bytes(micros(), &timeBytes[0]);  // send the time sent first
+  float2bytes((float) micros(), &timeBytes[0]);  // send the time sent first
   float2bytes(currentGPS->state.lat, &latBytes[0]);
   float2bytes(currentGPS->state.lon, &lonBytes[0]);
   sentBytes += BT_SERIAL.write(timeBytes, 4)
